logo

+++ app/Http/Controllers/UserController.php
@@ -119,7 +119,9 @@ class UserController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
-        Storage::disk('public')->delete($user->photo);
+        if ($user->photo) {
+            Storage::disk('public')->delete($user->photo);
+        }
         $user->delete();
         return redirect()->route('users.index')->with('success', 'User successfully deleted');
     }
